@@ -1,4 +1,14 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-import "./assets/global.css";
-createApp(App).mount("#app");
+import './index.css';
+
+import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+
+import App from './App.vue'
+import router from './router'
+
+const app = createApp(App)
+
+app.use(createPinia())
+app.use(router)
+
+app.mount('#app')
